util.js created
