+++ index.html
@@ -9,7 +9,8 @@
     <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap" rel="stylesheet">
     <title>Boaters Social</title>
     <link rel="stylesheet" href="dist/css/styles.css">
-    <script src="js/login.js"></script>
+    <script type="module" src="js/login.js"></script>
+    <script type="module" src="js/util.js"></script>
 </head>
 <body>
     <div class="card position-absolute top-50 start-50 translate-middle col-10 col-md-4">
