+++ js/util.js
@@ -1,0 +1,11 @@
+export const API_BASE_URL = 'https://api.noroff.dev';
+
+
+export function formatDate(dateString) {
+    const date = new Date(dateString);
+    const day = date.getUTCDate().toString().padStart(2, '0');
+    const month = (date.getUTCMonth() + 1).toString().padStart(2, '0');
+    const year = date.getUTCFullYear().toString();
+    return `${day}.${month}.${year}`;
+  }
+
