+++ profile/index.html
@@ -83,7 +83,8 @@
         </div>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="../js/feed.js"></script>
-    <script src="../js/profile.js"></script>
+    <script type="module" src="../js/feed.js"></script>
+    <script type="module" src="../js/profile.js"></script>
+    <script type="module" src="../js/util.js"></script>
 </body>
 </html>
